==> 1_CreacionLimpiezaDataset.ipynb <==
!pip3 install tqdm==4.62.0
!pip3 install pandas==1.3.1
!pip3 install numpy==1.21.1

import pandas as pd
import numpy as np
import glob
import json
import time
from tqdm import tqdm

#parent dir for all the dataset
input_dir = 'G:/BertV1/2022-01-03'

#metadata
metadata_path = f'{input_dir}/metadata.csv'
metadata = pd.read_csv(metadata_path,
                               dtype={'pubmed_id': str,'title': str,'abstract': str})
metadata.head()

#fetch the number of reasearch papers in pdf json folder
all_json_pdf = glob.glob(f'{input_dir}/document_parses/pdf_json/*.json',recursive=True)
len(all_json_pdf)

#fetch the number of reasearch papers in pmc json folder
all_json_pmc = glob.glob(f'{input_dir}/document_parses/pmc_json/*.json',recursive=True)
len(all_json_pmc)

#Class to extract paper id, abstract and body from research papers
class FileReader:
    def __init__(self, file_path):
        with open(file_path) as file:
            content = json.load(file)
            self.paper_id = content['paper_id']
            self.abstract = "$$$".join([each['text']  for each in content.get('abstract', [])])
            self.body_text = "$$$".join([each['text']  for each in content['body_text']])
    def __repr__(self):
        return f'{self.paper_id}:\nabstract:  {self.abstract[:200]}\nbody:  {self.body_text[:200]}...'

#showing a sample research paper from pdf json folder
first_row = FileReader(all_json_pdf[0])
print(first_row)

#Create a dictionary of all research papers from pdf json
dict_ = {'paper_id': [], 'abstract': [],'body_text': []}
t1 = time.time()
for idx, entry in tqdm(enumerate(all_json_pdf)):
    try:
        content = FileReader(entry)
        dict_['paper_id'].append(content.paper_id)
        dict_['abstract'].append(content.abstract)
        dict_['body_text'].append(content.body_text)
    except:
        print("An exception occurred")
print(time.time() - t1)

#Creating a dataframe of all research papers from pdf json
df_json_pdf = pd.DataFrame(dict_, columns=['paper_id', 'abstract','body_text'])
df_json_pdf.head()

#Create a dictionary of all research papers from pmc json
dict_ = {'paper_id': [],'body_text': []}
t1 = time.time()
for idx, entry in tqdm(enumerate(all_json_pmc)):
    content = FileReader(entry)
    dict_['paper_id'].append(content.paper_id)
    dict_['body_text'].append(content.body_text)
print(time.time() - t1)

#Creating a dataframe of all research papers from pmc json
df_json_pmc = pd.DataFrame(dict_, columns=['paper_id','body_text'])
df_json_pmc.head()

Muestreo y almacenamiento del conjunto de datos del curso

#drop rows from the metadata where the corresponding research text doesn't exist in pdf json
metadata.dropna(subset=['pdf_json_files'], inplace=True)

#Taking 25000 randmly sampled records from metadata
#sub_metadata = metadata.sample(25000)
sub_metadata = metadata.sample(50000)

#Sample the both pdf and pmc research paper table based on the sampled metadata table
sub_df_json_pdf = df_json_pdf[df_json_pdf['paper_id'].isin(sub_metadata['sha'])]
sub_df_json_pmc = df_json_pmc[df_json_pmc['paper_id'].isin(sub_metadata['pmcid'])]

#storing the sample data (used for our course)
sub_metadata.to_pickle('metadata_sample2.pickle')
sub_df_json_pdf.to_pickle('json_pdf_sample2.pickle')
sub_df_json_pmc.to_pickle('json_pmc_sample2.pickle')
