==> 2_PreprocesamientoDatos.ipynb <==
!pip3 install pandas==1.3.1
!pip3 install numpy==1.21.1

#for colab users and python<3.8
!pip3 install pickle5
import pickle5 as pickle

#for python >= 3.8 RUN below command and comment above command
#import pickle

import pandas as pd
import numpy as np

# Accessing the research papers and data

with open('C:/Users/Alexis/Desktop/SistemaQABert/metadata_sample2.pickle', 'rb') as f:
  df_metadata = pickle.load(f)
with open('C:/Users/Alexis/Desktop/SistemaQABert/json_pdf_sample2.pickle', 'rb') as f:
  df_pdf = pickle.load(f)
with open('C:/Users/Alexis/Desktop/SistemaQABert/json_pmc_sample2.pickle', 'rb') as f:
  df_pmc = pickle.load(f)

Muestreo y almacenamiento del conjunto de datos del curso


#Merging the metadata and json pdf data
#sha column of metadata and paper_id of json_pdf data are same identifier for each record
df_merged = pd.merge(df_metadata,df_pdf,left_on='sha',right_on='paper_id',how='left').drop('paper_id',axis=1)

df_merged.head(1)
#abstract_x is the abstract from metadata and abstract_y is the abstract from json_pdf

#Lets merge the json_pmc data to the merged data too
df_merged = pd.merge(df_merged,df_pmc,left_on='pmcid',right_on='paper_id',how='left').drop('paper_id',axis=1)
df_merged.shape

Limpieza y preprocesamiento de datos

df_merged[df_merged.abstract_x != df_merged.abstract_y].shape

# check metadata abstract column to see if null values exist
df_merged.abstract_x.isnull().sum(),(df_merged.abstract_x == '').sum()

# Check pdf_json abstract to see if null values exist
df_merged.abstract_y.isnull().sum(),(df_merged.abstract_y == '').sum()

Dado que el abstract_x de los metadatos es más fiable, lo utilizaremos pero sólo lo rellenaremos con el texto abstract_y cuando el valor de abstract_x sea nulo


# Convert all columns to string and then replace abstract_y values
#df = df.astype(str)
df_merged["abstract_y"] = df_merged["abstract_y"].astype(str) 
df_merged['abstract_y'] = np.where(df_merged['abstract_y'].map(len) > 50, df_merged['abstract_y'], "na")

df_merged[df_merged['abstract_y'].apply(lambda x: len(str(x)) <= 10)]

# sustituir abstract_x (columna de metadatos) por abstract_y (pdf_json) valor donde abstract_x es null
df_merged.loc[df_merged.abstract_x.isnull() & (df_merged.abstract_y != 'na'),'abstract_x'] = df_merged[df_merged.abstract_x.isnull() & (df_merged.abstract_y != 'na')].abstract_y

# ¿Tenemos algún valor abstracto nulo restante? Ya no. Esto es bueno.
# Los valores nulos se han reducido que es lo que esperábamos.
df_merged.abstract_x.isnull().sum()


# Deshagámonos de la columna pdf_json abstract y cambiemos el nombre de la columna metadata abstract
df_merged.rename(columns = {'abstract_x' : 'abstract'}, inplace = True)
df_merged.drop('abstract_y',axis=1,inplace = True)
df_merged.columns

# Esto se espera porque el texto del cuerpo viene de las carpetas pdf y pmc
(df_merged.body_text_x != df_merged.body_text_y).sum()

df_merged.body_text_x.isnull().sum(),(df_merged.body_text_y == '').sum()

# Esto se espera porque hay menos papeles en json_pmc
df_merged.body_text_y.isnull().sum()

# body_text_x es pdf_json. body_text_y viene de pmc_json
# Cuando está disponible utilizamos el texto del archivo pmc confiando en la calidad de la declaración
df_merged.body_text_x.isnull().sum(),(df_merged.body_text_y.isnull()).sum()

# Hay ~13k filas en las que body_text_x es null pero body_text_y no es null# Hay ~13k filas en las que body_text_x es null pero body_text_y no es null
df_merged.loc[df_merged.body_text_x.isnull() & df_merged.body_text_y.notnull()]

# Confiamos en que el texto de la carpeta pmc sea de mayor calidad ya que contiene el texto completo. 
# Por lo tanto, vamos a reemplazar con body_text_x con body_text_y donde body_text_y existe
df_merged.loc[df_merged.body_text_y.notnull(),'body_text_x'] = df_merged.loc[df_merged.body_text_y.notnull(), 'body_text_y']

# Deshagámonos de la columna pdf_pmc body text y cambiemos el nombre de la columna body textdf_merged.rename(columns = {'body_text_x' : 'body_text'}, inplace = True)
df_merged.rename(columns = {'body_text_x' : 'body_text'}, inplace = True)
df_merged.drop('body_text_y',axis=1,inplace = True)
df_merged.columns

# Los valores nulos del cuerpo del texto ahora han disminuido. 
df_merged.body_text.isnull().sum()

df_merged.columns

df_final = df_merged[['sha', 'title', 'abstract', 'publish_time', 'authors', 'url', 'body_text']]

df_final.head()

df_final = df_final.dropna(axis=0,subset=['body_text', 'title'])
df_final.isnull().sum()

df_final.shape

# GUARDANDO DATOS FILTRADOS

df_final.to_csv('FINAL_CORD_DATA2.csv', index=False)

 # Análisis Exploratorio de Datos - Data Analysis (EDA) 

!pip install seaborn==0.11.1

import seaborn as sns
sns.set_theme(style="darkgrid")
import calendar

Revisión de los datos finales 

#Inspect data
def inspect_data(data):
    return pd.DataFrame(
        {
            "Data Type": data.dtypes,
            "No of levels": data.apply(lambda x:len(x.unique()), axis=0),
            "levels": data.apply(lambda x:x.unique(), axis=0),
            "qtd_null": data.apply(lambda x: x.isnull().sum(), axis=0),
            "pct_null": data.apply(lambda x: x.isnull().sum() * 100 / len(x))
        }
    )

inspect_data(df_final)

df_final['publish_time'] = pd.to_datetime(df_final['publish_time'])

df_final['publish_year'] =  df_final['publish_time'].dt.year
df_final['publish_month'] =  df_final['publish_time'].dt.month
df_final['publish_day'] =  df_final['publish_time'].dt.day

df_final['publish_year'].nunique()

df_final['publish_year'].min()

df_final['publish_year'].max()

# Nº de artículos publicados en los últimos años
sns.countplot(x="publish_year", data=df_final[df_final['publish_year'].isin([2018, 2019,2020,2021, 2022])])

#Distribución por meses de los trabajos publicados
sns.countplot(x="publish_month", data=df_final)

df_final['publish_month'] = df_final['publish_month'].apply(lambda x : calendar.month_abbr[x])

sns.countplot(x="publish_month", data=df_final)

#Comprobación del % de trabajos publicados en tiempos recientes frente a años anteriores
df_final['in_covid_time'] = np.where(df_final['publish_year'].isin([2019,2020,2021]), 1, 0)
sns.countplot(df_final['in_covid_time'])

**Revisión de nubes del palabras**

!pip3 install wordcloud

from wordcloud import WordCloud, STOPWORDS 
import re
import string
import matplotlib.pyplot as plt
%matplotlib inline

def plot_word_cloud(wordcloud):
    plt.figure(figsize=(10, 10))
    plt.imshow(wordcloud) 
    plt.axis("off");

def remove_special_charac_and_punctuations(text) :
# define el patrón a mantener
  pattern = r'[^a-zA-z0-9\'\s]' 
  txt = re.sub(pattern, '', text)
  txt = ''.join([c for c in txt if c not in string.punctuation])
  return  txt

def concat_all_sentences(sents):
   all_tokens = ''
   for text in sents:
        tx = str(text).lower().strip()
        tx = " ".join(tx.split())
        all_tokens += tx + " "
   return all_tokens  

cleaned_titles = df_final.loc[df_final.title.notnull(), 'title'].apply(remove_special_charac_and_punctuations)

cleaned_abstracts = df_final.loc[df_final.abstract.notnull(), 'abstract'].apply(remove_special_charac_and_punctuations)

title_tokens = concat_all_sentences(cleaned_titles)

abstract_tokens = concat_all_sentences(cleaned_abstracts)

title_tokens

abstract_tokens

# Observando la nube de palabras de los tokens del título
title_wordcloud = WordCloud(width = 500, height = 500,random_state=10).generate(title_tokens)
plot_word_cloud(title_wordcloud)

## Modelización de temas

!pip install gensim==3.8.3

!pip install spacy

!python -m spacy download en_core_web_sm

import gensim, spacy, logging, warnings
import gensim.corpora as corpora
from gensim.utils import lemmatize, simple_preprocess
from gensim.models import CoherenceModel

abstracts = cleaned_abstracts.values.tolist()

# simple_preprocess of genism: elimina los caracteres especiales, los números y los convierte en minúsculas
gensim.utils.simple_preprocess("12 happy today @# '' '\n'", deacc=True) 

def token_builder(sentences) :
   for sent in sentences:     
        sent = gensim.utils.simple_preprocess(str(sent), deacc=True) 
        yield(sent)  


# Construcción de tokens abstractos usando token_builder
abstract_tokens = list(token_builder(abstracts))

# Construcción de los modelos de bigramas y trigramas
bigram = gensim.models.Phrases(abstract_tokens, min_count=3, threshold=50) # higher threshold fewer phrases.
trigram = gensim.models.Phrases(bigram[abstract_tokens], threshold=50)  
bigram_mod = gensim.models.phrases.Phraser(bigram)
trigram_mod = gensim.models.phrases.Phraser(trigram)

!pip install nltk

import nltk as nltk
nltk.download('stopwords')
nltk.download('punkt')
from nltk.corpus import stopwords

stop_words = stopwords.words('english')
stop_words.extend(['from', 'subject', 're', 'edu', 'use', 'not', 'would', 'say', 'could', '_', 'be', 'know', 'good', 'go', 'get', 'do', 'done', 'try', 'many', 'some', 'nice', 'thank', 'think', 'see', 'rather', 'easy', 'easily', 'lot', 'lack', 'make', 'want', 'seem', 'run', 'need', 'even', 'right', 'line', 'even', 'also', 'may', 'take', 'come'])

def process_docs(texts, stop_words=stop_words, allowed_postags=['NOUN', 'ADJ', 'VERB', 'ADV']):
    texts = [[word for word in simple_preprocess(str(doc)) if word not in stop_words] for doc in texts]
    texts = [bigram_mod[doc] for doc in texts]
    texts = [trigram_mod[bigram_mod[doc]] for doc in texts]
    texts_out = []
    # nlp = spacy.load('en', disable=['parser', 'ner'])
    nlp = spacy.load("en_core_web_sm")
    for sent in texts:
        doc = nlp(" ".join(sent)) 
        texts_out.append([token.lemma_ for token in doc if token.pos_ in allowed_postags])
    # Eliminar las palabras de parada una vez más después de la lematización   
    texts_out = [[word for word in simple_preprocess(str(doc)) if word not in stop_words] for doc in texts_out] 
    return texts_out

abstracts_data = process_docs(abstract_tokens)

# Creación de un Diccionario de datos
id2word = corpora.Dictionary(abstracts_data)

print(id2word)

print(id2word.token2id)

# Creación del Corpus: Term Document Frequency
corpus = [id2word.doc2bow(text) for text in abstracts_data]

# Construcción del modelo LDA 
lda_model = gensim.models.ldamodel.LdaModel(corpus=corpus,
                                           id2word=id2word,
                                           num_topics=2, 
                                           random_state=100,
                                           update_every=1,
                                           chunksize=10,
                                           passes=10,
                                           alpha='symmetric',
                                           iterations=100,
                                           per_word_topics=True)

print(lda_model.print_topics())

import matplotlib.pyplot as plt

import matplotlib.colors as mcolors

cols = [color for name, color in mcolors.TABLEAU_COLORS.items()]  # more colors: 'mcolors.XKCD_COLORS'

cloud = WordCloud(stopwords=stop_words,
                  background_color='white',
                  width=2500,
                  height=1800,
                  max_words=10,
                  colormap='tab10',
                  color_func=lambda *args, **kwargs: cols[i],
                  prefer_horizontal=1.0)

topics = lda_model.show_topics(formatted=False)

for i in range(0,2):
    topic_words = dict(topics[i][1])
    cloud.generate_from_frequencies(topic_words, max_font_size=300)
    plt.gca().imshow(cloud)
    plt.gca().set_title('Topic ' + str(i), fontdict=dict(size=16))
    plt.gca().axis('off')
    plt.show()

topics = lda_model.show_topics(formatted=False)

topics
